Move shared-steps out of `specs` into `lib`
