+++ lib/shared-steps/wp-signup-spec.js
@@ -1,7 +1,7 @@
 /** @format */
 import assert from 'assert';
-import { getBrowserSingleton } from '../../lib/driver-manager';
-import SitePreviewComponent from '../../lib/components/site-preview-component.js';
+import { getBrowserSingleton } from '../driver-manager';
+import SitePreviewComponent from '../components/site-preview-component.js';
 
 export const stepWrapper = ( { name, body } ) => {
 	step( name, async () => {
